main() is shrunk
- admittedly the complexity is merely moved to a subsidiary
  struct without changes. No tests have been added.
- this lays the groundwork for adding tests
